fix: https://github.com/pandas-dev/pandas/pull/62191#discussion_r2437358484
